Move UnsafeCode tests to pretty-tests (except for those that are not yet successful).
