+++ ICSharpCode.Decompiler.Tests/TestCases/Correctness/UnsafeCode.cs
@@ -22,7 +22,7 @@ namespace ICSharpCode.Decompiler.Tests.TestCases.Correctness
 {
 	public class UnsafeCode
 	{
-		struct SimpleStruct
+		private struct SimpleStruct
 		{
 			public int X;
 			public double Y;
@@ -32,66 +32,6 @@ namespace ICSharpCode.Decompiler.Tests.TestCases.Correctness
 		{
 			// TODO: test behavior, or convert this into a pretty-test
 			// (but for now, it's already valuable knowing whether the decompiled code can be re-compiled)
-		}
-
-		public unsafe int* NullPointer
-		{
-			get {
-				return null;
-			}
-		}
-
-		public unsafe int* PointerCast(long* p)
-		{
-			return (int*)p;
-		}
-
-		public unsafe long ConvertDoubleToLong(double d)
-		{
-			return *(long*)(&d);
-		}
-
-		public unsafe double ConvertLongToDouble(long d)
-		{
-			return *(double*)(&d);
-		}
-
-		public unsafe int ConvertFloatToInt(float d)
-		{
-			return *(int*)(&d);
-		}
-
-		public unsafe float ConvertIntToFloat(int d)
-		{
-			return *(float*)(&d);
-		}
-
-		public unsafe int PointerCasts()
-		{
-			int i = 0;
-			*(float*)&i = 0.5f;
-			((byte*)&i)[3] = 3;
-			return i;
-		}
-
-		public unsafe void PassRefParameterAsPointer(ref int p)
-		{
-			fixed (int* ptr = &p) {
-				this.PassPointerAsRefParameter(ptr);
-			}
-		}
-
-		public unsafe void PassPointerAsRefParameter(int* p)
-		{
-			this.PassRefParameterAsPointer(ref *p);
-		}
-
-		public unsafe void AddressInMultiDimensionalArray(double[,] matrix)
-		{
-			fixed (double* ptr = &matrix[1, 2]) {
-				this.PointerReferenceExpression(ptr);
-				this.PointerReferenceExpression(ptr);
-			}
 		}
 
 		public unsafe int MultipleExitsOutOfFixedBlock(int[] arr)
@@ -109,42 +49,7 @@ namespace ICSharpCode.Decompiler.Tests.TestCases.Correctness
 			Console.WriteLine("outside");
 			return 2;
 		}
-
-		public unsafe void FixedStringAccess(string text)
-		{
-			fixed (char* ptr = text) {
-				char* ptr2 = ptr;
-				while (*ptr2 != '\0') {
-					*ptr2 = 'A';
-					ptr2++;
-				}
-			}
-		}
-
-		public unsafe void PutDoubleIntoLongArray1(long[] array, int index, double val)
-		{
-			fixed (long* ptr = array) {
-				((double*)ptr)[index] = val;
-			}
-		}
-
-		public unsafe void PutDoubleIntoLongArray2(long[] array, int index, double val)
-		{
-			fixed (long* ptr = &array[index]) {
-				*(double*)ptr = val;
-			}
-		}
-
-		public unsafe string PointerReferenceExpression(double* d)
-		{
-			return d->ToString();
-		}
-
-		public unsafe string PointerReferenceExpression2(long addr)
-		{
-			return ((int*)addr)->ToString();
-		}
-
+		
 		public unsafe void FixMultipleStrings(string text)
 		{
 			fixed (char* ptr = text, userName = Environment.UserName, ptr2 = text) {
@@ -154,6 +59,11 @@ namespace ICSharpCode.Decompiler.Tests.TestCases.Correctness
 			}
 		}
 
+		public unsafe string UsePointer(double* ptr)
+		{
+			return ptr->ToString();
+		}
+
 		public unsafe string StackAlloc(int count)
 		{
 			char* ptr = stackalloc char[count];
@@ -161,64 +71,19 @@ namespace ICSharpCode.Decompiler.Tests.TestCases.Correctness
 			for (int i = 0; i < count; i++) {
 				ptr[i] = (char)i;
 			}
-			return this.PointerReferenceExpression((double*)ptr);
+			return this.UsePointer((double*)ptr);
 		}
 
 		public unsafe string StackAllocStruct(int count)
 		{
 			SimpleStruct* s = stackalloc SimpleStruct[checked(count * 2)];
 			SimpleStruct* p = stackalloc SimpleStruct[10];
-			return this.PointerReferenceExpression(&s->Y);
+			return this.UsePointer(&s->Y);
 		}
-
-		public unsafe int* PointerArithmetic(int* p)
-		{
-			return p + 2;
-		}
-
+		
 		public unsafe byte* PointerArithmetic2(long* p, int y, int x)
 		{
 			return (byte*)((short*)p + (y * x));
-		}
-
-		public unsafe long* PointerArithmetic3(long* p)
-		{
-			return (long*)((byte*)p + 3);
-		}
-
-		public unsafe long* PointerArithmetic4(void* p)
-		{
-			return (long*)((byte*)p + 3);
-		}
-
-		public unsafe int PointerArithmetic5(void* p, byte* q, int i)
-		{
-			return (int)(q[i] + *(byte*)p);
-		}
-
-		public unsafe int PointerSubtraction(long* p, long* q)
-		{
-			return (int)(p - q);
-		}
-
-		public unsafe long PointerSubtractionLong(long* p, long* q)
-		{
-			return p - q;
-		}
-
-		public unsafe int PointerSubtraction2(long* p, short* q)
-		{
-			return (int)((byte*)p - (byte*)q);
-		}
-
-		public unsafe int PointerSubtraction3(void* p, void* q)
-		{
-			return (int)((byte*)p - (byte*)q);
-		}
-
-		unsafe ~UnsafeCode()
-		{
-			this.PassPointerAsRefParameter(this.NullPointer);
 		}
 	}
 }
