+++ ICSharpCode.Decompiler.Tests/TestCases/ILPretty/.gitignore
@@ -1,0 +1,1 @@
+/*.dll
