+++ ICSharpCode.Decompiler.Tests/TestCases/Pretty/.gitignore
@@ -1,3 +1,4 @@
 /*.res
 /*.dll
 /*.exe
+/*.pdb
