+++ ICSharpCode.Decompiler/IL/ControlFlow/DetectPinnedRegions.cs
@@ -337,6 +337,7 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 					new PointerType(((ByReferenceType)oldVar.Type).ElementType),
 					oldVar.Index);
 				newVar.Name = oldVar.Name;
+				newVar.HasGeneratedName = oldVar.HasGeneratedName;
 				oldVar.Function.Variables.Add(newVar);
 				ReplacePinnedVar(oldVar, newVar, pinnedRegion);
 			} else if (pinnedRegion.Variable.Type.IsKnownType(KnownTypeCode.String)) {
@@ -383,6 +384,7 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 					if (nativeVar.Kind == VariableKind.Local) {
 						newVar = new ILVariable(VariableKind.PinnedLocal, nativeVar.Type, nativeVar.Index);
 						newVar.Name = nativeVar.Name;
+						newVar.HasGeneratedName = nativeVar.HasGeneratedName;
 						nativeVar.Function.Variables.Add(newVar);
 						ReplacePinnedVar(nativeVar, newVar, pinnedRegion);
 					} else {
